Renamed library folder to modules...
Just renamed a folder to make more sense and test something with Github projects.
