+++ BLEND-ArmatureRiggingModules/modules/chains/_spline_.py
